Move code common to both hermes_cli and hermes_sdk to a common module
